requirements.txt
streamlit==1.31.1
langchain==0.1.11
langchain-community==0.0.27
langchain-core==0.1.30
langchain-openai==0.0.8
langchain-text-splitters==0.0.1
langchain-openai==0.0.8
python-dotenv==1.0.1
grep==0.3.2
